backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
google-genai
supabase
python-dotenv
pydantic
